Combine preview and upload to deduplicate code.
