+++ backend/config/routes/api_station.php
@@ -296,11 +296,9 @@ return static function (RouteCollectorProxy $group) {
                                     $group->get('/bulk', Controller\Api\Stations\BulkMedia\DownloadAction::class)
                                         ->setName('api:stations:files:bulk');
 
-                                    $group->post('/bulk', Controller\Api\Stations\BulkMedia\UploadAction::class);
-
                                     $group->post(
-                                        '/bulk/preview',
-                                        Controller\Api\Stations\BulkMedia\PreviewAction::class
+                                        '/bulk[/{preview:preview}]',
+                                        Controller\Api\Stations\BulkMedia\UploadAction::class
                                     );
 
                                     $group->get('/download', Controller\Api\Stations\Files\DownloadAction::class)
