+++ backend/src/Controller/Api/Stations/BulkMedia/UploadAction.php
@@ -7,6 +7,7 @@ namespace App\Controller\Api\Stations\BulkMedia;
 use App\Container\EntityManagerAwareTrait;
 use App\Controller\SingleActionInterface;
 use App\Entity\Api\StationPlaylistImportResult;
+use App\Entity\Api\StationPlaylistPreviewResult;
 use App\Entity\Repository\CustomFieldRepository;
 use App\Entity\Repository\StationPlaylistMediaRepository;
 use App\Entity\Repository\StationPlaylistRepository;
@@ -19,6 +20,7 @@ use App\Http\Response;
 use App\Http\ServerRequest;
 use App\OpenApi;
 use App\Service\Flow;
+use App\Utilities\Types;
 use League\Csv\Reader;
 use OpenApi\Attributes as OA;
 use Psr\Http\Message\ResponseInterface;
@@ -35,6 +37,22 @@ use function str_starts_with;
         path: '/station/{station_id}/files/bulk',
         operationId: 'postStationBulkMediaUpload',
         summary: 'Upload a CSV containing details about all station media.',
+        tags: [OpenApi::TAG_STATIONS_MEDIA],
+        parameters: [
+            new OA\Parameter(ref: OpenApi::REF_STATION_ID_REQUIRED),
+        ],
+        responses: [
+            // TODO: API Response Body
+            new OpenApi\Response\Success(),
+            new OpenApi\Response\AccessDenied(),
+            new OpenApi\Response\NotFound(),
+            new OpenApi\Response\GenericError(),
+        ]
+    ),
+    OA\Post(
+        path: '/station/{station_id}/files/bulk/preview',
+        operationId: 'postStationBulkMediaPreview',
+        summary: 'Preview changes from a CSV containing details about all station media.',
         tags: [OpenApi::TAG_STATIONS_MEDIA],
         parameters: [
             new OA\Parameter(ref: OpenApi::REF_STATION_ID_REQUIRED),
@@ -83,15 +101,19 @@ final class UploadAction implements SingleActionInterface
             return $flowResponse;
         }
 
+        $isPreview = Types::string($params['preview'] ?? '') === 'preview';
+
         // Lookup tables for later.
         $mediaStorage = $station->media_storage_location;
 
         $mediaByPath = [];
         $mediaByUniqueId = [];
+        $mediaLookup = [];
 
         $mediaInStorageLocation = $this->em->createQuery(
             <<<DQL
-            SELECT sm.id, sm.unique_id, sm.path
+            SELECT sm.id, sm.unique_id, sm.path, sm.title, sm.artist,
+                   sm.album, sm.genre, sm.lyrics, sm.isrc, sm.extra_metadata_raw
             FROM App\Entity\StationMedia sm
             WHERE sm.storage_location = :storageLocation
             DQL
@@ -101,6 +123,7 @@ final class UploadAction implements SingleActionInterface
         foreach ($mediaInStorageLocation as $mediaRow) {
             $mediaByPath[md5($mediaRow['path'])] = $mediaRow['id'];
             $mediaByUniqueId[$mediaRow['unique_id']] = $mediaRow['id'];
+            $mediaLookup[$mediaRow['id']] = $mediaRow;
         }
 
         $extraMetadataFieldNames = StationMediaMetadata::getFields();
@@ -122,8 +145,10 @@ final class UploadAction implements SingleActionInterface
         $reader = Reader::createFromPath($csvPath);
         $reader->setHeaderOffset(0);
 
-        $processed = 0;
         $importResults = [];
+        $previewResults = [];
+        $totalChanges = 0;
+        $processed = 0;
 
         $i = 0;
         $batchSize = 50;
@@ -138,41 +163,80 @@ final class UploadAction implements SingleActionInterface
                 continue;
             }
 
-            $record = $this->em->find(StationMedia::class, $mediaId);
-            if (!($record instanceof StationMedia)) {
-                continue;
-            }
-
-            unset($row['id'], $row['path']);
-
-            $importResult = [
-                'id' => $record->id,
-                'title' => $record->title,
-                'artist' => $record->artist,
-                'success' => false,
-                'error' => null,
-            ];
-
-            try {
-                $rowResult = $this->processRow(
-                    $record,
-                    $station,
-                    $row,
-                    $extraMetadataFieldNames,
-                    $customFieldShortNames,
-                    $playlistsByName
-                );
-
-                $importResult['success'] = $rowResult;
-                if ($rowResult) {
-                    $processed++;
+            if ($isPreview) {
+                if (!isset($mediaLookup[$mediaId])) {
+                    continue;
                 }
-            } catch (Throwable $e) {
-                $importResult['success'] = false;
-                $importResult['error'] = $e->getMessage();
-            }
 
-            $importResults[] = $importResult;
+                $currentMedia = $mediaLookup[$mediaId];
+                unset($row['id'], $row['path']);
+
+                $previewResult = [
+                    'id' => $mediaId,
+                    'title' => $currentMedia['title'],
+                    'artist' => $currentMedia['artist'],
+                    'has_changes' => false,
+                    'changes' => [],
+                    'error' => null,
+                ];
+
+                try {
+                    $changes = $this->previewRow(
+                        $currentMedia,
+                        $station,
+                        $row,
+                        $extraMetadataFieldNames,
+                        $customFieldShortNames,
+                        $playlistsByName
+                    );
+
+                    if (!empty($changes)) {
+                        $previewResult['has_changes'] = true;
+                        $previewResult['changes'] = $changes;
+                        $totalChanges++;
+                    }
+                } catch (Throwable $e) {
+                    $previewResult['error'] = $e->getMessage();
+                }
+
+                $previewResults[] = $previewResult;
+            } else {
+                $record = $this->em->find(StationMedia::class, $mediaId);
+                if (!($record instanceof StationMedia)) {
+                    continue;
+                }
+
+                unset($row['id'], $row['path']);
+
+                $importResult = [
+                    'id' => $record->id,
+                    'title' => $record->title,
+                    'artist' => $record->artist,
+                    'success' => false,
+                    'error' => null,
+                ];
+
+                try {
+                    $rowResult = $this->processRow(
+                        $record,
+                        $station,
+                        $row,
+                        $extraMetadataFieldNames,
+                        $customFieldShortNames,
+                        $playlistsByName
+                    );
+
+                    $importResult['success'] = $rowResult;
+                    if ($rowResult) {
+                        $processed++;
+                    }
+                } catch (Throwable $e) {
+                    $importResult['success'] = false;
+                    $importResult['error'] = $e->getMessage();
+                }
+
+                $importResults[] = $importResult;
+            }
 
             $i++;
             if (0 === $i % $batchSize) {
@@ -183,6 +247,16 @@ final class UploadAction implements SingleActionInterface
         $this->clearMemory();
 
         @unlink($csvPath);
+
+        if ($isPreview) {
+            return $response->withJson(
+                new StationPlaylistPreviewResult(
+                    message: sprintf(__('%d files will be modified.'), $totalChanges),
+                    previewResults: $previewResults,
+                    totalChanges: $totalChanges,
+                )
+            );
+        }
 
         return $response->withJson(
             new StationPlaylistImportResult(
@@ -279,6 +353,161 @@ final class UploadAction implements SingleActionInterface
         }
 
         return true;
+    }
+
+    private function previewRow(
+        array $currentMedia,
+        Station $station,
+        array $row,
+        array $extraMetadataFieldNames,
+        array $customFieldShortNames,
+        array $playlistsByName
+    ): array {
+        $changes = [];
+        $mediaChanges = [];
+        $extraMetadata = [];
+
+        $hasCustomFields = false;
+        $customFields = [];
+
+        $hasPlaylists = false;
+        $playlists = [];
+
+        foreach ($row as $key => $value) {
+            if ('' === $value) {
+                $value = null;
+            }
+
+            if (in_array($key, self::ALLOWED_MEDIA_FIELDS, true)) {
+                $currentValue = $currentMedia[$key] ?? null;
+                if ($currentValue !== $value) {
+                    $mediaChanges[$key] = [
+                        'field' => $key,
+                        'current' => $currentValue,
+                        'new' => $value,
+                    ];
+                }
+            } elseif (in_array($key, $extraMetadataFieldNames, true)) {
+                $currentExtraMetadata = $currentMedia['extra_metadata_raw'] ?? [];
+                $currentValue = $currentExtraMetadata[$key] ?? null;
+                if ($currentValue !== $value) {
+                    $extraMetadata[$key] = [
+                        'field' => $key,
+                        'current' => $currentValue,
+                        'new' => $value,
+                    ];
+                }
+            } elseif (str_starts_with($key, 'custom_field_')) {
+                $fieldName = str_replace('custom_field_', '', $key);
+                if (isset($customFieldShortNames[$fieldName])) {
+                    $hasCustomFields = true;
+                    $customFields[$customFieldShortNames[$fieldName]] = $value;
+                }
+            } elseif ('playlists' === $key) {
+                $hasPlaylists = true;
+                if (null !== $value) {
+                    foreach (explode(', ', $value) as $playlistName) {
+                        $playlistShortName = StationPlaylist::generateShortName($playlistName);
+                        if (isset($playlistsByName[$playlistShortName])) {
+                            /** @var int $playlistId */
+                            $playlistId = $playlistsByName[$playlistShortName];
+                            $playlists[$playlistId] = 0;
+                        }
+                    }
+                }
+            }
+        }
+
+        if (!empty($mediaChanges)) {
+            $changes['metadata'] = array_values($mediaChanges);
+        }
+
+        if (!empty($extraMetadata)) {
+            $changes['extra_metadata'] = array_values($extraMetadata);
+        }
+
+        if ($hasPlaylists) {
+            // Get current playlists for this media
+            $currentPlaylists = $this->em->createQuery(
+                <<<DQL
+                SELECT sp.id, sp.name
+                FROM App\Entity\StationPlaylistMedia spm
+                JOIN spm.playlist sp
+                WHERE spm.media = :mediaId
+                DQL
+            )->setParameter('mediaId', $currentMedia['id'])
+                ->getArrayResult();
+
+            $currentPlaylistIds = array_column($currentPlaylists, 'id');
+            $newPlaylistIds = array_keys($playlists);
+
+            $addedPlaylists = array_diff($newPlaylistIds, $currentPlaylistIds);
+            $removedPlaylists = array_diff($currentPlaylistIds, $newPlaylistIds);
+
+            if (!empty($addedPlaylists) || !empty($removedPlaylists)) {
+                $playlistChanges = [];
+
+                if (!empty($addedPlaylists)) {
+                    $addedPlaylistNames = [];
+                    foreach ($playlistsByName as $name => $id) {
+                        if (in_array($id, $addedPlaylists, true)) {
+                            $addedPlaylistNames[] = $name;
+                        }
+                    }
+                    $playlistChanges['added'] = $addedPlaylistNames;
+                }
+
+                if (!empty($removedPlaylists)) {
+                    $removedPlaylistNames = [];
+                    foreach ($currentPlaylists as $playlist) {
+                        if (in_array($playlist['id'], $removedPlaylists, true)) {
+                            $removedPlaylistNames[] = $playlist['name'];
+                        }
+                    }
+                    $playlistChanges['removed'] = $removedPlaylistNames;
+                }
+
+                $changes['playlists'] = $playlistChanges;
+            }
+        }
+
+        if ($hasCustomFields) {
+            // Get current custom fields for this media
+            $currentCustomFields = $this->em->createQuery(
+                <<<DQL
+                SELECT cf.short_name, smcf.value
+                FROM App\Entity\StationMediaCustomField smcf
+                JOIN smcf.field cf
+                WHERE smcf.media = :mediaId
+                DQL
+            )->setParameter('mediaId', $currentMedia['id'])
+                ->getArrayResult();
+
+            $currentCustomFieldValues = [];
+            foreach ($currentCustomFields as $field) {
+                $currentCustomFieldValues[$field['short_name']] = $field['value'];
+            }
+
+            $customFieldChanges = [];
+            foreach ($customFields as $fieldId => $newValue) {
+                $fieldShortName = array_search($fieldId, $customFieldShortNames, true);
+                $currentValue = $currentCustomFieldValues[$fieldShortName] ?? null;
+
+                if ($currentValue !== $newValue) {
+                    $customFieldChanges[] = [
+                        'field' => $fieldShortName,
+                        'current' => $currentValue,
+                        'new' => $newValue,
+                    ];
+                }
+            }
+
+            if (!empty($customFieldChanges)) {
+                $changes['custom_fields'] = $customFieldChanges;
+            }
+        }
+
+        return $changes;
     }
 
     private function clearMemory(): void
